Fixed up creator attribute on appointments in order to manage permissions for editing and deleting them.

+++ backend/src/main/java/de/hase/hasev2/appointment/AppointmentService.java
@@ -1,6 +1,9 @@
 package de.hase.hasev2.appointment;
 
+import de.hase.hasev2.appointment.exceptions.AppointmentNotFoundException;
+import de.hase.hasev2.appointment.exceptions.NotAppointmentCreatorException;
 import de.hase.hasev2.config.HikariService;
+import de.hase.hasev2.user.User;
 import org.jooq.DSLContext;
 import org.jooq.impl.DSL;
 import org.slf4j.Logger;
@@ -29,6 +32,13 @@ public class AppointmentService {
         }
     }
 
+    private boolean isCreator(int appointmentId, int userId) throws AppointmentNotFoundException {
+        var appointment = this.findAppointment(appointmentId)
+                .orElseThrow(() -> new AppointmentNotFoundException(appointmentId));
+
+        return appointment.creator() == userId;
+    }
+
     public List<Appointment> findAllAppointments() {
         return database.selectFrom(APPOINTMENTS)
                 .fetchInto(Appointment.class);
@@ -40,24 +50,33 @@ public class AppointmentService {
                 .fetchOptionalInto(Appointment.class);
     }
 
-    public Optional<Appointment> saveAppointment(Appointment appointment) {
-        return database.insertInto(APPOINTMENTS, APPOINTMENTS.NAME, APPOINTMENTS.DATE, APPOINTMENTS.LOCATION)
-                .values(appointment.name(), appointment.date(), appointment.location())
+    public Optional<Appointment> saveAppointment(Appointment appointment, User creator) {
+        return database.insertInto(APPOINTMENTS, APPOINTMENTS.NAME, APPOINTMENTS.CREATOR, APPOINTMENTS.DATE, APPOINTMENTS.LOCATION)
+                .values(appointment.name(), creator.matrikelNr(), appointment.date(), appointment.location())
                 .returningResult()
                 .fetchOptionalInto(Appointment.class);
     }
 
-    public Optional<Appointment> deleteAppointment(int appointmentId) {
+    public Optional<Appointment> deleteAppointment(int appointmentId, User creator) throws AppointmentNotFoundException, NotAppointmentCreatorException {
+        if(!this.isCreator(appointmentId, creator.matrikelNr())) {
+            throw new NotAppointmentCreatorException(appointmentId, creator.matrikelNr());
+        }
+
         return database.deleteFrom(APPOINTMENTS)
                 .where(APPOINTMENTS.APPOINTMENTID.eq(appointmentId))
                 .returningResult()
                 .fetchOptionalInto(Appointment.class);
     }
 
-    public Optional<Appointment> updateAppointment(Appointment updatedAppointment){
+    public Optional<Appointment> updateAppointment(Appointment updatedAppointment, User creator) throws AppointmentNotFoundException, NotAppointmentCreatorException {
+        if(!this.isCreator(updatedAppointment.appointmentId(), creator.matrikelNr())) {
+            throw new NotAppointmentCreatorException(updatedAppointment.appointmentId(), creator.matrikelNr());
+        }
+
         return database.update(APPOINTMENTS)
                 .set(APPOINTMENTS.DATE, updatedAppointment.date())
                 .set(APPOINTMENTS.NAME, updatedAppointment.name())
+                .set(APPOINTMENTS.CREATOR, updatedAppointment.creator())
                 .set(APPOINTMENTS.LOCATION, updatedAppointment.location())
                 .where(APPOINTMENTS.APPOINTMENTID.eq(updatedAppointment.appointmentId()))
                 .returningResult()
